requirements.txt
asgiref==3.7.2
Django==3.2.25
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.4.4
typing-extensions==4.7.1
